Decoder refactor - adding output projections
This commit introduces the output_projection parameter of a decoder.
The output projection is the function that takes the previousl RNN state,
last generated (embedded) output and a list of context vectors, and produces
a vector.

Outside this function, this vector is linearly projected onto the size
of the vocabulary, and then, softmax is computed over it.

There are three output projections:
- no_deep_output: returns concatenation of the inputs
- maxout_output: equivalent to deep output with one maxout hidden layer
- mlp_output: the concatenation of the inputs is projected through multilayer perceptron